Input Values Functional
Values 1 through 9 and the decimal point buttons are now functional to being displayed in by the text box properly.

The default "0" values is replaced unless a decimal point is added first.

Each numerical button press will append the value to the end of the string displayed in the text box field.

The clear button is functional in clearing the text box.