Basic Operations
Basic operations of addition, subtraction, multiplication and division implemented.

Currently there is a known problem when pressing the same operation multiple times in a row.

Equals button now computes the first value entered and the second value entered and sets the first value to be the output of the operation.

Clear Entry button has been implemented to simply clear the text box, values that are stored are not lost.

Clear button has been expanded to set values to be default values (0.0).